requirements/common.txt
boto == 2.38.0
carrot == 0.10.7
celery == 3.1.18
Django >= 1.7 , < 1.8
django-celery == 3.1.16
django_compressor == 1.5
django-contact-form == 1.0
django-extensions == 1.5.5
django-flatblocks == 0.9.1
django-floppyforms == 1.5.2
django-haystack == 2.4.0
django-mailer == 1.1
django-markup == 0.4
django-publicmanager == 0.9.2
django-reset == 0.2.0
django-ses == 0.7.0
django-sortedm2m == 1.0.2
django-taggit == 0.16.4
django-template-email == 0.1
django-templatequery == 0.3.0
django-tinymce == 2.0.4
ghettoq == 0.4.5
gunicorn == 19.3.0
pyparsing == 2.0.3
django-mediastore == 0.7.0
beautifulsoup4 == 4.4.0
http://pypi-81b384.ma-work.co.uk/media/dists/sorl-thumbnail-3.2.5-ma1.tar.gz
django-tastypie == 0.12.1